Lite-HRNet-SimCC/requirements.txt
matplotlib==3.5.1
numpy==1.23.4
opencv_python==4.5.5.64
pandas==1.2.4
Pillow==8.4.0
pycocotools==2.0.4
PyYAML==6.0
seaborn==0.11.2
thop==0.1.1.post2209072238
torch==1.11.0
torchvision==0.12.0
tqdm==4.61.2
